fstab-gen: use the smallest device timeout value when 'nofail' is set
When 'nofail' is set, we shouldn't wait for the device timeout to
expire. This was the case before but the way it was done was broken.

This patch restores this behaviour by setting the smallest timeout
value (1us) when 'nofail' is set.

This also applies to swap units since they use the same path to setup
device timeout.