Spare libxslt a bit of memory management burden by hoisting document() out of the inner function call and moving it up by a few levels so it's inside the determineCrossID template.
This ought not to make a difference for memory consumption,
because the documents never need to be kept open concurrently; except,
it turns out that it does because somehow calling document() triggers
a memory leak in libxslt or something.

This change may lop as much as 400MB off of the maximum rss size as
reported by /usr/bin/time -v on a make XSLT_DEBUG_CROSS_REFS=1 call.