nspawn: refuse to start a container out of it cgroup namespace root
When the systemd-nspawn is in a cgroup that is not reachable from the
current cgroup namespace, we cannot mount the cgroup for the container
in any meaningful way.

systemd or systemd-nspawn does not lead to this situation. However, the
user could use "unshare --cgroup" and then migrate the process outside
of its cgroup namespace's root, then run systemd-nspawn.