journal: coalesce ftruncate()s in 100ms windows
Prior to this change every journal append causes an ftruncate() for the
sake of inotify propagation of the mmap-based writes.

With this change the notification is deferred up to ~100ms, coalescing
any repeated journal writes during the deferred period into a single
ftruncate().  The ftruncate() call isn't free and doing it on every
append adds unnecessary overhead and latency in the journald event loop.

A bit of surgery was required in journal-file.c to provide a callback
interface for notification and quiescing of the concluding journal
during rotation.