journal-remote: Use monotonic time of receiver machine in log timestamp
Replace the monotonic timestamp of the log entry with the monotonic timestamp of the receiver machine. This is useful when the receiver receives logs from multiple machines. In this case it is not guaranteed that the monotonic timestamps of the arriving log entries will be increasing, as they come from multiple machines. This causes the journal file to continuously rotate, filling up the disk space very quickly.
In order to preserve all the information in the original log, the field _SOURCE_MONOTONIC_TIMESTAMP is appended to the log entry, with its value being the monotonic timestamp of the sender machine. For example, the following line is appended:
_SOURCE_MONOTONIC_TIMESTAMP=512142575
The field _SOURCE_MONOTONIC_TIMESTAMP is added only if it does not already exist, so that the original monotonic timestamp is preserved if the remote logs are forwarded to a different machine.